CORE-973 Add App launch info header and deprecation messages.
Added a toolbar containing the app name,
with the app description under it.

The app launch stepper form will be replaced with a help message
if the app is deleted, disabled, or uses any deprecated parameters.

Refactored most of the form into a new AppLaunchForm file.

Added a story for an app with a TreeSelection parameter
(the only deprecated parameter for now).
